% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_zero.R
\name{check_zero}
\alias{check_zero}
\title{Check zeros in a vector}
\usage{
check_zero(x, type = c("any", "all"))
}
\description{
check_zero(c(1,1), 'all')
check_zero(c(1,1), 'any')
check_zero(c(0,1), 'any')
check_zero(c(0,1), 'all')
check_zero(c(0,0), 'all')
}
